+++ allo/dataflow_opt/__init__.py
@@ -1,0 +1,1 @@
+from .passes import *
