+++ allo/dataflow_opt/passes.py
@@ -1,0 +1,315 @@
+from allo._mlir.ir import (
+    Location,
+    MemRefType,
+    InsertionPoint,
+    IntegerSetAttr,
+    Block,
+)
+from allo._mlir.dialects import (
+    func as func_d,
+    affine as affine_d,
+    memref as memref_d,
+)
+from allo._mlir.ir import StringAttr
+from allo._mlir.passmanager import PassManager as mlir_pass_manager
+from allo._mlir._mlir_libs._mlir.ir import Module
+from allo.customize import Schedule
+from allo.ir.transform import find_func_in_module
+
+def dataflow_optimization_pass(schedule: Schedule, debugPoint=None) -> Schedule:
+    fn_name = schedule.top_func.name.value
+    mod = _mlir_preprocess(schedule.module, fn_name)
+    assert check_valid_kernel(mod)
+
+    # Dataflow canonicalization pass 
+    mod = _dataflow_canonicalization_pass(mod)
+    if debugPoint == "dataflow_canonicalization":
+        return Schedule(
+            mod,
+            find_func_in_module(mod, fn_name),
+            schedule.func_args,
+            schedule.ip,
+            schedule.ext_libs,
+            schedule.inst_list,
+        )
+
+    # Other passes
+    return Schedule(
+        mod,
+        find_func_in_module(mod, fn_name),
+        schedule.func_args,
+        schedule.ip,
+        schedule.ext_libs,
+        schedule.inst_list,
+    )
+
+
+def check_valid_kernel(mod: Module) -> bool:
+    # unimplemented
+    return True
+
+def _mlir_preprocess(module, top_func_name):
+    """
+    Performs linalg-to-affine lowering, then aggressive inlining on an MLIR module, then removes all functions except the top-level function.
+    """
+    # Configure for maximum inlining - no recursion limit and always inline
+    MAX_ITER = INLINE_THRESHOLD = 999999
+    pipeline = f"builtin.module(convert-linalg-to-affine-loops,inline{{max-iterations={MAX_ITER} inlining-threshold={INLINE_THRESHOLD}}},symbol-privatize{{exclude={top_func_name}}},symbol-dce)"
+    try:
+        with module.context:
+            mlir_pass_manager.parse(pipeline).run(module.operation)
+        return module
+    except Exception as e:
+        print("Error: failed to run MLIR passes, printing module...")
+        print(module)
+        raise e
+    
+def _dataflow_canonicalization_pass(module):
+    """
+    Implements the dataflow canonicalization pass as described in the Stream-HLS paper (https://arxiv.org/pdf/2501.09118)
+
+    This pass ensures that the program is compatible with dataflow architectures by transforming shared buffers to adhere to single-producer-single-consumer patterns. This pass does not handle complex patterns involving multiple producers writing to the same buffer, except in the case of reduction loops.
+    """
+    with module.context, Location.unknown():
+        for op in module.body.operations:
+            if not isinstance(op, func_d.FuncOp):
+                continue
+            canonicalize_fn(op)
+    return module
+
+
+def canonicalize_fn(op: func_d.FuncOp):
+    ops = list(op.entry_block.operations)
+    for op_in_block in ops:
+        if isinstance(op_in_block, memref_d.AllocOp):
+            canonicalize_alloc(op_in_block)
+        elif isinstance(op_in_block, func_d.CallOp):
+            canonicalize_call(op_in_block)
+
+
+def canonicalize_call(op: func_d.CallOp):
+    for result in op.results:
+        loads = []  # (op, idx)
+        stores = []  # ops
+        for use in result.uses:
+            user = use.owner
+            if isinstance(
+                user, (memref_d.LoadOp, affine_d.AffineLoadOp, func_d.CallOp)
+            ):
+                for idx, operand in enumerate(user.operands):
+                    if operand == op.result:
+                        loads.append((user, idx))
+            elif isinstance(user, (memref_d.StoreOp, affine_d.AffineStoreOp)):
+                stores.append(user)
+
+        if not isinstance(result.type, MemRefType) or len(loads) == 1:
+            # already single produer single consumer pattern
+            continue
+
+        if len(stores) > 1:
+            raise NotImplementedError(
+                "Complex pattern detected in call op; additional canonicalization not implemented yet."
+            )
+
+        n_loads = len(loads)
+        ip = InsertionPoint(op)
+        new_allocs = [
+            memref_d.AllocOp(result.type, [], [], ip=ip) for _ in range(n_loads)
+        ]
+
+        shape = result.type.shape
+
+        loop_ivs, loops = [], []
+        loop_ip = ip
+        for dim in shape:
+            loop = affine_d.AffineForOp(
+                lower_bound=0, upper_bound=dim, step=1, ip=loop_ip
+            )
+            loop_ivs.append(loop.induction_variable)
+            loops.append(loop)
+            loop_ip = InsertionPoint(loop.body)
+
+        loops[0].move_after(op)
+
+        with InsertionPoint(loops[-1].body):
+            orig_value = affine_d.AffineLoadOp(
+                result.type.element_type,
+                result,
+                loop_ivs,
+                affine_d.AffineMap.get_identity(len(shape)),
+            )
+            for alloc in new_allocs:
+                affine_d.AffineStoreOp(
+                    orig_value.result,
+                    alloc,
+                    loop_ivs,
+                    affine_d.AffineMap.get_identity(len(shape)),
+                )
+
+        for loop in loops:
+            with InsertionPoint(loop.body):
+                affine_d.AffineYieldOp([])
+
+        for idx, (user, op_idx) in enumerate(loads):
+            user.operation.operands[op_idx] = new_allocs[idx].result
+
+
+def canonicalize_alloc(alloc_op):
+    loads = []  # (op, idx)
+    stores = []  # ops
+    returns = []  # ops
+    for use in alloc_op.result.uses:
+        user = use.owner
+        if isinstance(user, (memref_d.LoadOp, affine_d.AffineLoadOp, func_d.CallOp)):
+            for idx, operand in enumerate(user.operands):
+                if operand == alloc_op.result:
+                    loads.append((user, idx))
+        elif isinstance(user, (memref_d.StoreOp, affine_d.AffineStoreOp)):
+            stores.append(user)
+        elif isinstance(user, func_d.ReturnOp):
+            returns.append(user)
+
+    memref_type = alloc_op.result.type
+    shape = memref_type.shape
+    orig_name = (
+        alloc_op.attributes["name"].value if "name" in alloc_op.attributes else "buffer"
+    )
+    if len(shape) == 0:
+        # should constants be propogated?
+        return
+
+    # single store with multiple loads.
+    if len(stores) == 1 and len(loads) > 1:
+        store = stores[0]
+        for i, (load, idx) in enumerate(loads[1:]):
+            new_alloc = alloc_op.operation.clone(ip=InsertionPoint(alloc_op))
+            new_alloc.attributes["name"] = StringAttr.get(f"{orig_name}_split_{i}")
+            store_dup = store.clone(ip=InsertionPoint(store))
+            store_dup.operation.replace_uses_of_with(alloc_op.result, new_alloc.result)
+            load.operation.operands[idx] = new_alloc.result
+        return
+
+    # store-load-store-load loop redunction pattern
+    if len(stores) == 2 and len(loads) + len(returns) == 2:
+        l_ops = [l[0] for l in loads]
+        l_ops.extend(returns)
+        if store_load_store_load_pattern(alloc_op, l_ops, stores):
+            return
+
+    # multiple loads and multiple stores
+    if len(stores) >= 2 or len(loads) >= 2:
+        raise NotImplementedError(
+            "Complex pattern detected in alloc op; additional canonicalization not implemented yet."
+        )
+
+
+def store_load_store_load_pattern(alloc_op, loads, stores):
+    """
+    Transforms reduction loops to satisfy the condition that the number of writes to a shared buffer equals the number of reads.
+    """
+    assert len(loads) == 2 and len(stores) == 2
+
+    loop_load, loop_store = None, None
+
+    # find loop_load and loop_store
+    for load in loads:
+        for store in stores:
+            if load.parent == store.parent:
+                loop_load = load
+                loop_store = store
+                break
+        if loop_load:
+            break
+
+    if not loop_load or not loop_store:
+        return False
+
+    store_op = [s for s in stores if s != loop_store][0]
+    load_op = [l for l in loads if l != loop_load][0]
+
+    # check for unsupported store_op in an if block
+    parent = store_op.parent
+    while parent:
+        if isinstance(parent, affine_d.AffineIfOp):
+            return False
+        parent = parent.parent
+
+    loop_nest = []
+    current_op = loop_load.parent
+    while current_op:
+        if isinstance(current_op.opview, affine_d.AffineForOp):
+            loop_nest.append(current_op.opview)
+        current_op = current_op.parent
+    if not loop_nest:
+        return False
+
+    ip = InsertionPoint(alloc_op)
+    memref_type = alloc_op.result.type
+    new_alloc1 = memref_d.AllocOp(memref_type, [], [], ip=ip)
+    new_alloc2 = memref_d.AllocOp(memref_type, [], [], ip=ip)
+
+    loop_ivs = [loop.induction_variable for loop in loop_nest]  # innermost IV first
+
+    with InsertionPoint.at_block_begin(loop_nest[0].body):
+        first_iter_set = affine_d.IntegerSet.get(
+            1, 0, [affine_d.AffineExpr.get_dim(0)], [True]
+        )
+
+        first_iter_if = affine_d.AffineIfOp(
+            results_=[], _gen_arg_0=[loop_ivs[0]], loc=Location.unknown()
+        )
+
+        first_iter_if.attributes["condition"] = IntegerSetAttr.get(first_iter_set)
+
+    # In the if block, load from original buffer and store to new_alloc1
+    then_block = Block.create_at_start(parent=first_iter_if.thenRegion)
+    with InsertionPoint(then_block):
+        new_load = affine_d.AffineLoadOp(
+            memref_type.element_type, alloc_op.result, loop_load.indices, loop_load.map
+        )
+        affine_d.AffineStoreOp(
+            new_load.result, new_alloc1.result, loop_load.indices, loop_load.map
+        )
+        affine_d.AffineYieldOp([])
+
+    loop_load.operation.replace_uses_of_with(alloc_op.result, new_alloc1.result)
+    loop_store.operation.replace_uses_of_with(alloc_op.result, new_alloc1.result)
+
+    inner_loop_upper_map = loop_nest[-1].upperBoundMap.value
+
+    # check upper bound is a constant
+    if not (
+        inner_loop_upper_map.n_dims == 0 and len(inner_loop_upper_map.results) == 1
+    ):
+        return False
+
+    loop_upper_bound = inner_loop_upper_map.results[0]
+
+    last_iter_set = affine_d.IntegerSet.get(
+        1, 0, [affine_d.AffineExpr.get_dim(0) - loop_upper_bound + 1], [True]
+    )
+
+    last_iter_if = affine_d.AffineIfOp(
+        results_=[], _gen_arg_0=[loop_ivs[0]], loc=Location.unknown(), ip=ip
+    )
+
+    last_iter_if.move_after(loop_store)
+
+    last_iter_if.attributes["condition"] = IntegerSetAttr.get(last_iter_set)
+
+    final_then_block = Block.create_at_start(parent=last_iter_if.thenRegion)
+    with InsertionPoint(final_then_block):
+        final_loop_load = affine_d.AffineLoadOp(
+            memref_type.element_type,
+            new_alloc1.result,
+            loop_load.indices,
+            loop_load.map,
+        )
+        affine_d.AffineStoreOp(
+            final_loop_load.result, new_alloc2.result, loop_load.indices, loop_load.map
+        )
+        affine_d.AffineYieldOp([])
+
+    load_op.operation.replace_uses_of_with(alloc_op.result, new_alloc2.result)
+
+    return True
